refactor(services,resolver): another one bites the dust
Importer service in this case.
