+++ crates/services/importer/src/strong_app.rs
@@ -3,17 +3,18 @@ use std::{fs, sync::Arc};
 use async_graphql::Result;
 use chrono::{Duration, NaiveDateTime};
 use csv::ReaderBuilder;
-use itertools::Itertools;
-use models::{
-    DeployStrongAppImportInput, SetLot, UserExerciseInput, UserWorkoutInput, UserWorkoutSetRecord,
-    WorkoutSetStatistic,
+use dependent_models::ImportResult;
+use fitness_models::{
+    SetLot, UserExerciseInput, UserWorkoutInput, UserWorkoutSetRecord, WorkoutSetStatistic,
 };
+use itertools::Itertools;
+use media_models::DeployStrongAppImportInput;
 use regex::Regex;
 use rust_decimal::Decimal;
 use rust_decimal_macros::dec;
 use serde::{Deserialize, Serialize};
 
-use super::{app_utils, ImportResult};
+use super::utils;
 
 #[derive(Debug, Serialize, Deserialize, Clone, Default)]
 #[serde(rename_all = "PascalCase")]
@@ -107,7 +108,7 @@ pub async fn import(
         if next_entry.date != entry.date {
             let ndt = NaiveDateTime::parse_from_str(&entry.date, "%Y-%m-%d %H:%M:%S")
                 .expect("Failed to parse input string");
-            let ndt = app_utils::get_date_time_with_offset(ndt, timezone.clone());
+            let ndt = utils::get_date_time_with_offset(ndt, timezone.clone());
             let re = Regex::new(r"^(\d+h)?\s?(\d+m)?$").unwrap();
             let workout_duration = if let Some(captures) = re.captures(&entry.workout_duration) {
                 let hours = captures.get(1).map_or(0, |m| {
