+++ crates/services/importer/src/trakt.rs
@@ -1,8 +1,11 @@
+use application_utils::get_base_http_client;
 use async_graphql::Result;
+use common_utils::APPLICATION_JSON_HEADER;
 use convert_case::{Case, Casing};
+use dependent_models::ImportResult;
 use enums::{ImportSource, MediaLot, MediaSource};
 use itertools::Itertools;
-use models::{
+use media_models::{
     CreateOrUpdateCollectionInput, DeployTraktImportInput, ImportOrExportItemRating,
     ImportOrExportItemReview, ImportOrExportMediaItemSeen,
 };
@@ -11,9 +14,8 @@ use rust_decimal::Decimal;
 use rust_decimal_macros::dec;
 use sea_orm::prelude::DateTimeUtc;
 use serde::{Deserialize, Serialize};
-use utils::{get_base_http_client, APPLICATION_JSON_HEADER};
 
-use super::{ImportFailStep, ImportFailedItem, ImportOrExportMediaItem, ImportResult};
+use super::{ImportFailStep, ImportFailedItem, ImportOrExportMediaItem};
 
 const API_URL: &str = "https://api.trakt.tv";
 const CLIENT_ID: &str = "b3d93fd4c53d78d61b18e0f0bf7ad5153de323788dbc0be1a3627205a36e89f5";
